# feat!: improved performance for sim compilation