--- a/libs/wingsdk/src/core/app.ts
+++ b/libs/wingsdk/src/core/app.ts
@@ -62,9 +62,18 @@ export abstract class App extends Construct {
   }
 
   /**
-   * Directory where artifacts are synthesized to.
+   * The output directory.
    */
-  public abstract readonly workdir: string;
+  public abstract readonly outdir: string;
+
+  /**
+   * The ".wing" directory, which is where the compiler emits its output. We are taking an implicit
+   * assumption here that it is always set to be `$outdir/.wing` which is currently hard coded into
+   * the `cli/compile.ts` file.
+   */
+  public get workdir() {
+    return `${this.outdir}/.wing`;
+  }
 
   /**
    * Synthesize the app into an artifact.
@@ -146,18 +155,14 @@ export abstract class App extends Construct {
  */
 export abstract class CdktfApp extends App {
   /**
-   * Directory where artifacts are synthesized to.
-   */
-  public readonly workdir: string;
-  /**
    * Path to the Terraform manifest file.
    */
   public readonly terraformManifestPath: string;
+  public readonly outdir: string;
 
   private readonly cdktfApp: cdktf.App;
   private readonly cdktfStack: cdktf.TerraformStack;
   private readonly pluginManager: PluginManager;
-  private readonly outdir: string;
 
   private synthed: boolean;
   private synthedOutput: string | undefined;
@@ -174,6 +179,8 @@ export abstract class CdktfApp extends App {
     const cdktfStack = new cdktf.TerraformStack(cdktfApp, TERRAFORM_STACK_NAME);
 
     super(cdktfStack, "Default");
+
+    this.outdir = outdir;
 
     // HACK: monkey patch the `new` method on the cdktf app (which is the root of the tree) so that
     // we can intercept the creation of resources and replace them with our own.
@@ -195,7 +202,6 @@ export abstract class CdktfApp extends App {
     this.pluginManager = new PluginManager(props.plugins ?? []);
 
     this.outdir = outdir;
-    this.workdir = cdktfOutdir;
     this.cdktfApp = cdktfApp;
     this.cdktfStack = cdktfStack;
     this.terraformManifestPath = join(this.outdir, "main.tf.json");
@@ -267,6 +273,9 @@ export abstract class CdktfApp extends App {
 
     this.synthed = true;
     this.synthedOutput = stringify(cleaned, null, 2) ?? "";
+
+    // clean up working directory
+    rmSync(this.workdir, { recursive: true });
 
     return this.synthedOutput;
   }