--- a/libs/wingsdk/src/target-sim/app.ts
+++ b/libs/wingsdk/src/target-sim/app.ts
@@ -1,7 +1,6 @@
 import * as fs from "fs";
 import * as path from "path";
 import { Construct } from "constructs";
-import * as tar from "tar";
 import { Bucket } from "./bucket";
 import { Counter } from "./counter";
 import { Function } from "./function";
@@ -21,7 +20,7 @@ import {
 import { SDK_VERSION } from "../constants";
 import * as core from "../core";
 import { preSynthesizeAllConstructs } from "../core/app";
-import { mkdtemp, SIMULATOR_FILE_PATH } from "../util";
+import { SIMULATOR_FILE_PATH } from "../util";
 
 /**
  * A construct that knows how to synthesize simulator resources into a
@@ -29,27 +28,17 @@ import { mkdtemp, SIMULATOR_FILE_PATH } from "../util";
  */
 export class App extends core.App {
   /**
-   * Directory where artifacts are synthesized to.
-   */
-  public readonly workdir: string;
-
-  /**
    * The output directory of this app.
    */
-  protected readonly outdir: string;
+  public readonly outdir: string;
 
-  private readonly name: string;
-  private readonly simfile: string;
   private synthed = false;
 
   constructor(props: core.AppProps) {
     super(undefined as any, "root");
-    this.name = props.name ?? "app";
     this.outdir = props.outdir ?? ".";
-    this.workdir = mkdtemp();
 
     Logger.register(this);
-    this.simfile = path.join(this.outdir, `${this.name}.wsim`);
   }
 
   protected tryNew(
@@ -90,32 +79,21 @@ export class App extends core.App {
    */
   public synth(): string {
     if (this.synthed) {
-      return this.simfile;
+      return this.outdir;
     }
 
     // call preSynthesize() on every construct in the tree
     preSynthesizeAllConstructs(this);
 
     // write simulator.json file into workdir
-    this.synthSimulatorFile(this.workdir);
+    this.synthSimulatorFile(this.outdir);
 
     // write tree.json file into workdir
-    core.synthesizeTree(this, this.workdir);
-
-    // tar + gzip the workdir, and write it as a .wsim file to the simfile
-    tar.create(
-      {
-        gzip: true,
-        cwd: this.workdir,
-        sync: true,
-        file: this.simfile,
-      },
-      ["./"]
-    );
+    core.synthesizeTree(this, this.outdir);
 
     this.synthed = true;
 
-    return this.simfile;
+    return this.outdir;
   }
 
   private synthSimulatorFile(outdir: string) {