--- a/libs/wingsdk/src/testing/sim-app.ts
+++ b/libs/wingsdk/src/testing/sim-app.ts
@@ -23,7 +23,7 @@ export class SimApp extends sim.App {
   public async startSimulator(): Promise<Simulator> {
     this.synthIfNeeded();
     const simfile = this.synth();
-    const s = new Simulator({ simfile });
+    const s = new Simulator({ simdir: simfile });
     await s.start();
     return s;
   }