@MarkMcCulloh 2023-03-20
Would it not be useful for the simulator functions to be self-contained as well? I worry this creates an unnecessary behavior difference between sim and non-sim targets.