@eladb 2023-03-20
> Would it not be useful for the simulator functions to be self-contained as well?

We can talk about created self-contained `.wsim` file (in which case maybe the right approach would be to simply bundle the npm dependencies instead of bundling each function separately). But that's not P1 right now. We are obsessively optimizing for dev iteration time.

> I worry this creates an unnecessary behavior difference between sim and non-sim targets.

There's a huge difference between sim and non-sim targets in almost every aspect. Same as with every other aspect of our SDK, we will need good test coverage to guarantee that our resources behave the same across the targets. As the project matures, we will get to the point where we have a good approximation. 

Does that make sense?