@eladb 2023-03-20
> Added a label so this PR gets a minor semver bump

Shouldn't that happen automatically when we specify `BREAKING CHANGE`?