@Chriscbr 2023-03-20
> I worry this creates an unnecessary behavior difference between sim and non-sim targets.

Yeah, I agree this change does potentially increase the gap between targets (specifically, things like module resolution are always a little bit sensitive / error prone), but I think it's probably outweighed by the improvement this makes for iteration loops while working with the Wing simulator.

Personally I also will miss our short-lived `.wsim` file format 😢 - but I think packaging simulated Wing apps is always something we can add later (as a separate feature) if our users really want it.