@eladb 2023-03-21
> Personally I also will miss our short-lived .wsim file format

Me too. It will come back with a vengeance!